Split Partition.hs out from CompileIoT.hs
This is largely self-contained and conceptually separate from the
rest of the stuff in CompileIoT. Split it out as an aid to testing
and comprehension.
