+++ striot.cabal
@@ -29,6 +29,7 @@ library
                       Striot.VizGraph
                       Striot.Jackson
                       Striot.Simple
+                      Striot.Partition
   -- other-modules:
   -- other-extensions:
   build-depends:      base              >= 4.9
@@ -109,4 +110,5 @@ test-suite test-striot
                       Striot.VizGraph
                       Striot.Jackson
                       Striot.Simple
+                      Striot.Partition
   default-language:   Haskell2010
